pyproject.toml
[project]
name = "zero-access-encryption"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "granian[reload,uvloop]>=2.7.2",
    "orjson>=3.11.7",
    "prek>=0.3.8",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "redis>=7.4.0",
    "rich>=14.3.3",
    "scalar-fastapi>=1.8.1",
    "sqlalchemy[asyncio]>=2.0.48",
    "sqlmodel>=0.0.37",
    "srp>=1.0.22",
    "structlog>=25.5.0",
]

[tool.pytest.ini_options]
testpaths = [ "src/tests" ]
pythonpath = [ "." ]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
lint.select = ["I"]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.0",
    "prek>=0.3.8",
    "pytest-asyncio>=0.24.0",
    "pytest>=9.0.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.8",
    "ty>=0.0.26",
    "fakeredis>=2.34.1",
]
